@@ -4,21 +4,21 @@ publish_to: "none" # Remove this line if you wish to publish to pub.dev
 version: 1.0.0+1
 
 environment:
-  sdk: ">=2.16.0 <3.0.0"
+  sdk: ">=3.0.0 <4.0.0"
 
 dependencies:
   flutter:
     sdk: flutter
-  cupertino_icons: ^1.0.2
+  cupertino_icons: ^1.0.8
 
 dev_dependencies:
-  emulators: ^0.4.8
+  emulators: ^0.5.1
   flutter_driver:
     sdk: flutter
-  flutter_lints: ^1.0.4
+  flutter_lints: ^4.0.0
   flutter_test:
     sdk: flutter
-  test: ^1.16.8
+  test: ^1.25.2
 
 flutter:
   uses-material-design: true
